faq-publish-api/requirements.txt
celery==5.0.5
Flask==1.1.2
Flask-Cors==3.0.10
flask-swagger-ui==3.36.0
names==0.3.0
nltk==3.5
requests==2.25.0
ruamel.yaml==0.16.12
SQLAlchemy==1.3.20
python-dotenv==0.15.0
psycopg2-binary==2.8.6